requirements.txt
accelerate==0.21.0
einops==0.7.0
ema-pytorch==0.2.3
numpy==1.25.1
opencv-python==4.5.5.64
Pillow==9.5.0
pytorch-lightning==1.4.2
torchvision==0.12.0
tqdm==4.65.0
scikit-learn==1.3.2
pandas==2.0.3
